Added Elite ship data
These are shamelessly copied from this github repository:
https://github.com/markmoxon/elite-beebasm

Due to the lack of a license, I probably don't have permission to do
anything with this data. So if I ever get asked to stop putting these
files here, I will have to remove this commit, and replace the ship models
with something else.

Signed-off-by: David Jander <djander@gmail.com>