cgb.py: Add support for color areas
Can adjust foreground and background colors if character cells. This
implies that color is shared by an area of 2x4 pixels.
To speed up color processing, the best way to use this is to setup the
color map (color information for rectangular regions of the screen)
beforehand and then draw by only setting or clearing pixels.

Signed-off-by: David Jander <djander@gmail.com>