cbg.py: implement clipped_line() method
This is not yet a full line clipping method, but for now only contains
basic checks to see is any part of the line is visible at all. This already
can have a huge impact on performance. Needs investigating if further
implementation of a line clipping algorithm has additional benefits.
If lines endpoints are perfectly clipped to the clipping area, then the
clipping check in putpixel can be optimized out. Until then... here's this.

Signed-off-by: David Jander <djander@gmail.com>